requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage
firebase_admin
python-docx
pandas
